pyproject.toml
[project]
name = "backend-template"
version = "0.1.0"
description = ""
authors = [
    {name = "RNT13",email = "renatornt13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "faker (>=37.12.0,<38.0.0)",
    "django-extensions (>=4.1,<5.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^25.9.0"
isort = "^7.0.0"
flake8 = "^7.3.0"
